Text.Numerals.Algorithm

Description

A module that contains data types and functions to automatically convert a number to words. It has tooling for a NumeralsAlgorithm as well as a HighNumberAlgorithm that is used to generate a ShortScale or LongScale.

Data types for number algorithms

data NumeralsAlgorithm

A data type for algorithmic number to word conversions. Most western languages likely can work with this data type.

numeralsAlgorithm :: (Foldable f, Foldable g) => Text -> Text -> Text -> f Text -> g (Integer, Text) -> FreeValueSplitter -> FreeMergerFunction -> (Text -> Text) -> FreeNumberToWords -> ClockText -> NumeralsAlgorithm

A smart constructor for the NumeralsAlgorithm type. This constructor allows one to use an arbitrary Foldable type for the low words and mid words. It will also order the midwords accordingly.

Large number algorithms

data HighNumberAlgorithm

A data type used for to map larger numbers to words. This data type supports the short scale and long scale with Latin prefixes, and custom suffixes. The Default value is the short scale with illion as suffix. This is used in English for large numbers.

Constructors

ShortScale Text

LongScale Text Text

shortScale :: Text -> FreeValueSplitter

Construct a FreeValueSplitter function for the given suffix for a short scale.

longScale :: Text -> Text -> FreeValueSplitter

Construct a FreeValueSplitter function for the given suffixes for a long scale.

shortScaleTitle :: Text -> FreeValueSplitter

Construct a FreeValueSplitter function for the given suffix for a short scale, the names are written in title case.

longScaleTitle :: Text -> Text -> FreeValueSplitter

Construct a FreeValueSplitter function for the given suffixes for a long scale, the names are written in title case.

valueSplit'

Arguments

:: (Text -> Text)
The post-processing function.

-> HighNumberAlgorithm
The HighNumberAlgorithm that is used.

-> FreeValueSplitter
The FreeValueSplitter result.

Generate a value splitter for a HighNumberAlgorithm but where the result is post-processed by a function.

Conversion to a NumberSegment

toSegments

Arguments

:: Integral i

=> Vector Text
A Vector of low words.

-> [(Integer, Text)]
The list of name and the names of these numbers in descending order for the mid words.

-> ValueSplitter i
The ValueSplitter used for large numbers, likely a splitter from a short scale or long scale.

-> NumberSegmenting i
The function that maps the number to the NumberSegment.

Convert the given number to a NumberSegment with the given Vector of low numbers, the sorted list of mid numbers, and a FreeValueSplitter for large numbers.

toSegmentLow

Arguments

:: Integral i

=> Vector Text
A Vector of low words.

-> NumberSegmenting i
The function that maps the number to the NumberSegment.

Convert the given number to a NumberSegment with the given Vector of low numbers. Mid words and large numbers are not taken into account. This is often the next step after the toSegmentMid.

toSegmentMid

Arguments

:: Integral i

=> Vector Text
A Vector of low words.

-> [(Integer, Text)]
The list of name and the names of these numbers in descending order for the mid words.

-> NumberSegmenting i
The function that maps the number to the NumberSegment.

Convert the given number to a NumberSegment with the given Vector of low numbers, and the sorted list of mid numbers. Large numbers are not taken into account. This is often the next step after the toSegmentHigh.

toSegmentHigh

Arguments

:: Integral i

=> Vector Text
A Vector of low words.

-> [(Integer, Text)]
The list of name and the names of these numbers in descending order for the mid words.

-> ValueSplitter i
The ValueSplitter used for large numbers, likely a splitter from a short scale or long scale.

-> NumberSegmenting i
The function that maps the number to the NumberSegment.

Convert the given number to a NumberSegment with the given Vector of low numbers, the sorted list of mid numbers, and a FreeValueSplitter for large numbers.

Segment compression

compressSegments

Arguments

:: Integral i

=> Text
The value used for one in the specific language.

-> MergerFunction i
The MergerFunction for the specific language that implements the grammar rules how to merge values.

-> NumberSegment i
The given NumberSegment value to turn into a Text object.

-> Text
The Text object that contains the name of the number stored in the NumberSegment.

Use the given MergerFunction to compress the NumberSegment to a single Text object that represents the given number.