Text.Numerals.Languages

Description

This module imports the num to word algorithms and re-exports these algorithms. The module thus can be used to conveniently import algorithms for all supported languages.

Documentation

dutch

Arguments

:: NumeralsAlgorithm
A NumeralsAlgorithm that can be used to convert numbers to different formats.

A NumeralsAlgorithm to convert numbers to words in the Dutch language.

english

Arguments

:: NumeralsAlgorithm
A NumeralsAlgorithm that can be used to convert numbers to different formats.

A NumeralsAlgorithm to convert numbers to words in the English language.

french

Arguments

:: NumeralsAlgorithm
A NumeralsAlgorithm that can be used to convert numbers to different formats.

A NumeralsAlgorithm to convert numbers to words in the French language.

german

Arguments

:: NumeralsAlgorithm
A NumeralsAlgorithm that can be used to convert numbers to different formats.

A NumeralsAlgorithm to convert numbers to words in the German language.