Text.Numerals.Languages.English

Description

This module contains logic to convert numbers to words in the English language.

Num to word algorithm

english

Arguments

:: NumeralsAlgorithm
A NumeralsAlgorithm that can be used to convert numbers to different formats.

A NumeralsAlgorithm to convert numbers to words in the English language.

Convert a cardinal number to text

toCardinal'

Arguments

:: Integral i

=> i
The number to convert to text.

-> Text
The cardinal counterpart in English.

Convert numers to their cardinal counterpart in English.

Convert to ordinal

ordinize' :: Text -> Text

A function that converts a number in words in cardinal form to ordinal form according to the English language rules.

Constant words

negativeWord' :: Text

The words used to mark a negative number in the English language.

zeroWord' :: Text

The word used for the number zero in the English language.

oneWord' :: Text

The word used for the number one in the English language.

Names for numbers

lowWords' :: Vector Text

A Vector that contains the word used for the numbers two to twenty in the English language.

midWords' :: [(Integer, Text)]

A list of 2-tuples that contains the names of values between thirty and thousand in the English language.

highWords' :: HighNumberAlgorithm

An algorithm to obtain the names of large numbers (one million or larger) in English. English uses a short scale with the illion suffix.

Merge function

merge' :: Integral i => i -> i -> Text -> Text -> Text

A merge function that is used to combine the names of words together to larger words, according to the English grammar rules.

Orphan instances

Default NumeralsAlgorithm

Instance details

Methods

def :: NumeralsAlgorithm