Text.Numerals.Prefix

Description

A module that defines Latin prefixes. These prefixes are used to construct names for the long and short scales. So the m, b, tr in million, billion, trillion.

Latin prefixes

latinPrefixes

Arguments

:: Vector Text
A Vector of Text objects to allow fast lookup.

The Latin prefixes in a Vector for O(1) lookup.

latinPrefixes'

Arguments

:: [Text]
A list of Text objects. This makes explicit recursion more convenient.

A list of Latin prefixes, used for the long and short scale.

latinPrefix

Arguments

:: Integral i

=> i
The value to map on a Latin prefix.

-> Maybe Text
The corresponding Latin prefix, given this exists.

Lookup the given Latin prefix for the given value.